Tablebases as a guide in self-play
The approach to the use of tablebases has been re-thought. A successful tablebase probe now, instead of immediately terminating self-play, *guides* self-play, so that optimal endgame play data is accumulated during training. Whereas the task of checking for a successful probe previously lay with the the `ChessEnv` class--frankly, constituting a violating encapsulation--it now resides within the `ChessPlayer`'s `action` method. In short, in cases where a tablebase probe is possible, the MCTS is simply _bypassed_, and a "fake" policy vector is created in which the optimal move is highlighted.